Extracted CodemirrorEditorSupport module to separate file
